requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
Django==1.10.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.1.0
Unidecode==0.04.19
certifi==2015.11.20.1
django-appconf==1.0.1
django-autoslug==1.9.3
django-browserid==2.0.0
django-cache-machine==0.9.1
django-compressor==2.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-haystack==2.4.1
django-jinja==2.2.1
django-taggit==0.18.0
elasticsearch==1.9.0
oauthlib==1.0.3
psycopg2==2.6.1
pyelasticsearch==1.4
python-dateutil==2.4.2
python-oauth2==1.0.0
python-twitter==2.2
pytz==2015.7
rcssmin==1.0.6
requests==2.9.1
requests-oauthlib==0.6.0
rjsmin==1.0.12
simplejson==3.8.1
six==1.10.0
smartypants==1.8.6
sorl-thumbnail==12.4a1
sqlparse==0.1.18
typogrify==2.0.7
unicode-slugify==0.1.3
urllib3==1.14
wheel==0.26.0
whitenoise==2.0.6
boto==2.42.0
django-storages==1.5.1

# For user registration, either via email or social
# Well-built with regular release cycles!
#django-allauth==0.24.1
